Move FavIconUtils to Site module
